// ts/svelte/free2z/src/lib/api/f2z.schemas.ts
export interface Event {
  readonly id: string;
  type: TypeEnum;
  /**
   * @minimum -2147483648
   * @maximum 2147483647
   * @nullable
   */
  amount?: number | null;
  readonly receiving_user: SimpleCreator;
  readonly contributor: SimpleCreator;
  readonly payee: SimpleCreator;
  readonly message: string;
  readonly content_url: string;
  readonly created_at: string;
  readonly read: boolean;
}

export interface FeaturedImage {
  readonly id: number;
  readonly url: string;
  readonly name: string;
  readonly thumbnail: string;
  readonly created_at: string;
  readonly updated_at: string;
  /** @maxLength 128 */
  title?: string;
  /** @maxLength 255 */
  description?: string;
  access?: AccessEnum;
  mime_type: MimeTypeEnum;
}
// ...
export interface PaginatedEventList {
  count: number;
  /** @nullable */
  next?: string | null;
  /** @nullable */
  previous?: string | null;
  results: Event[];
}
// ...
export interface SimpleCreator {
  /**
   * Required. 150 characters or fewer. Letters, digits and @/./+/-/_ only.
   * @maxLength 150
   * @pattern ^[\w.@+-]+$
   */
  username: string;
  avatar_image: FeaturedImage;
  /** @maxLength 255 */
  p2paddr?: string;
  /**
   * Price for 30 days of membership, livestream access, etc
   * @nullable
   * @pattern ^-?\d{0,6}(?:\.\d{0,0})?$
   */
  member_price?: string | null;
}
// ...
export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TypeEnum = {
  AISCORE_COMMENT: 'AISCORE_COMMENT',
  MADE_COMMENT: 'MADE_COMMENT',
  VOTE_COMMENT_RECEIVED: 'VOTE_COMMENT_RECEIVED',
  VOTE_COMMENT_SENT: 'VOTE_COMMENT_SENT',
  NEW_REPLY: 'NEW_REPLY',
  PAGE_UPDATE: 'PAGE_UPDATE',
  PAGE_VOTE: 'PAGE_VOTE',
  DONATION_RECEIVED: 'DONATION_RECEIVED',
  DONATION_SENT: 'DONATION_SENT',
  SUBSCRIPTION_RECEIVED: 'SUBSCRIPTION_RECEIVED',
  SUBSCRIPTION_SENT: 'SUBSCRIPTION_SENT',
  BOUGHT_TUZI: 'BOUGHT_TUZI',
  CHAT2Z: 'CHAT2Z',
  PLATFORM_FEE: 'PLATFORM_FEE',
  PAY_PER_VIEW_REVENUE: 'PAY_PER_VIEW_REVENUE',
  PAY_PER_VIEW_SPEND: 'PAY_PER_VIEW_SPEND',
} as const;